excel report

+++ BuisnessLogic/BindingModels/EmployeeBindingModel.cs
@@ -14,7 +14,7 @@ namespace BuisnessLogic.BindingModels
         public string EmployeeMiddlename { get; set; }
         public string EmployeePhoneNumber { get; set; }
         public string EmployeePassword { get; set; }
-        public int? EmployeePrize { get; set; }
+        public decimal? EmployeePrize { get; set; }
 
     }
 }
