+++ BuisnessLogic/BindingModels/WorkBindingModel.cs
@@ -10,6 +10,6 @@ namespace BuisnessLogic.BindingModels
     {
         public int? Id { get; set; }
         public string WorkName { get; set; }
-        public int WorkPrice { get; set; }
+        public decimal WorkPrice { get; set; }
     }
 }
