+++ BuisnessLogic/ViewModels/RepairViewModel.cs
@@ -9,7 +9,7 @@ namespace BuisnessLogic.ViewModels
         public string Name { get; set; }
         public DateTime DateStart { get; set; }
         public DateTime? DateEnd { get; set; }
-        public int Sum { get; set; }
+        public decimal Sum { get; set; }
         public int? ClientId { get; set; }
         public int? EmployeeId { get; set; }
         public int? WorkId { get; set; }
