+++ DatabaseImplement/Models/Employee.cs
@@ -23,7 +23,7 @@ namespace DatabaseImplement.Models
         [Required]
         public string EmployeePassword { get; set; }
         [Required]
-        public int EmployeePrize { get; set; }
+        public decimal EmployeePrize { get; set; }
         [ForeignKey("StaffId")]
         public virtual List<Staff> Staffs { get; set; }
         public virtual List<Repair> Repairs { get; set; }
